Send command as path instead of param
Not sure about this.  The advantage is that there are fewer params, but what do
you do for actions that take no arguments, like `stop`?  That becomes, e.g.:

    curl http://localhost:5000/stop

which curl will treat as a GET request.  Is that appropriate?  Actions with no
parameters can be GETs?  Or is that confusing/inconsistent?

We can force all requests to be a POST:

    curl -X POST http://localhost:5000/stop

but then we're not posting any data, so what's the point?

The Rdio web service API, for example, sends all requests through
https://services.rdio.com/api/1/ with the method sent as a parameter.